Develop

+++ lib/quantization.hpp
@@ -39,6 +39,6 @@ constexpr float qmatrix[2][64] = {
 void create_scaled_qtable(int c, int QF, int16_t *qtable);
 namespace jpegenc_hwy {
 namespace HWY_NAMESPACE {
-HWY_ATTR void quantize_core(int16_t *HWY_RESTRICT data, const int *HWY_RESTRICT qtable);
+HWY_ATTR void quantize_core(int16_t *HWY_RESTRICT data, const int16_t *HWY_RESTRICT qtable);
 }  // namespace HWY_NAMESPACE
 }  // namespace jpegenc_hwy
